feat: app registry + developer-signed settlement, local x402 facilitator, sparse training gradients, pq-qc AIR, batch/zk-quorum consensus, full SVM
- node: permissionless app registry with developer-signed settlement
  (tenzro_registerApp/settleAuthorized) — node never holds the
  payment-provider secret; streaming re-prefill failover on proxied
  provider drop
- payments: self-hosted local x402 facilitator as the default settlement
  route (operator's own EVM signer, no remote facilitator endpoint)
- training: chunked top-k + 2-bit sparse gradient codec with
  error-feedback residual carry-forward
- zk: PqQcAggregationAir (circuit_id "pq-qc") for ML-DSA-65 quorum-cert
  aggregation
- consensus: batch certificates and quorum-gated ZK commitment attestation
- vm: real SVM execution under the svm-full feature
- docs: align crate/root/SDK READMEs with the above; NOTICE + REFERENCES.md

+++ crates/tenzro-bridge/src/axelar.rs
@@ -444,7 +444,7 @@ impl BridgeAdapter for AxelarAdapter {
         Ok(BridgeTokenReceipt::new(
             format!("0x{}", hex::encode(hash.as_bytes())),
             hash,
-            (Timestamp::now().as_secs() * 1_000) as i64,
+            Timestamp::now().as_secs() * 1_000,
             0,
             request.source_chain,
             canonical_dest,
